bug: Correctly specify order for steps in recipe
Originally the `recipe::step_nominal()` was applied to the `train` data _before_ `recipe::step_normalize()`. This meant
that factor variables were converted to `0`/`1` which are _numeric_ and because in the call to
`recipe::step_normalize()` we used the argument `all_numeric()` it then normalised the newly created dummy variables.

This is obvious and expected behaviour...in hindsight!.

This commit therefore changes the order in which data is normalised (first and using `recipes::all_numeric_predictors()`
to succinctly and explicitly only normalised the predictors) and _then_ generate the dummy variables (but using the
argument `recipes::all_nominal_predictors()` rather than explicitly stating them, this makes it more flexible in the
future).

Another small change has been made to the `r/shf_thy_nod.R` which simplifies the conversion of character variables to
factor variables.