chore: Correcty tpyos and tidy logistic regression
There are two tpyo's in the CSV column headers that I've corrected as I keep on typing the correct spelling and getting
errors.

+ `rapid_enlargment` > `rapid_enlargement`
+ `compressive_symtoms` > `compressive_symptoms`

@mdp21oe : You should correct these in the copy of the `data/csv/sheffield_thyroid_nodule.csv` so that the script will
run correctly.

I've also split the logistic regression into its own section and rather than the manual extraction of coefficients and
confidence intervals and exponentiating them and combining them into a single table I've used
`gtsummary::tbl_regression()` to summarise the coefficients. This swaps one set of fiddly writing for another but is
perhaps a little easier to work with.