chore: Subsections correction; gtsummary regression
Having subsections in the `sections/` directory was causing problems when rendering as Quarto treats each child `.qmd`
as an individual "Notebook" and attempts to render it when building a manuscript. These Notebooks couldn't access
dataframes or objects created in `index.qmd`.

The solution was to prefix each with an `_` which forces them to be ignored as Notebooks. Each is then included verbatim
via the `{{< include sections/_file.qmd >}}` lines from `index.qmd`.

I've also moved the `setup` and `recipe` code chunk to their own child documents and collated all of the logistic
regression sections into their own child document which is included under the results section (for now).

To make life easy I've also switched to using `gtsummary::tbl_regression()` to summarise the results of these logistic
regression models.