imputation: Improves imputation layout and summary tables
Imputation is tweaked and the suggested table that summarises the data under each method and imputation is now
included. In doing so I've removed some of the duplicated tables that were starting to grow in favour of these imputed
tables.

Code is left in place as it may be desirable to use some of them in say an Appendix.