feature(imputation): Function to wrap mice and ggmice
Define a function that is a wrapper around imputation using `mice` and plotting the imputation paths and resulting
values using `ggmice`.

Not quite perfect legends of the bar charts are an issue that needs addressing but this is something that can be
addressed in due course as a separate issue (to be created).

Extract the legends from individual plots and add them to the end of each row, see the [cowplot shared
+legends article](https://wilkelab.org/cowplot/articles/shared_legends.html) for pointers on how to do this. Should
+ideally also get the `fill` colours to align with those used by `ggmice`.