improved function for tidymodelling
The function now works fairly well, different model specifications are built and a single call to fit the models with
the workflow is made. The results are summarised with importance tables and plots being extracted and then we fit the
model to the test cohort that was left out of model fitting. Metrics are calculated on this cohort and ROC plot
generated.